backend/requirements.txt
django>=4.2,<5
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
cloudinary>=1.36,<2
django-cloudinary-storage>=0.3,<1
whitenoise>=6.5,<7
dj-database-url>=2.1,<3
django-cors-headers>=4.3,<5
psycopg2-binary>=2.9,<3
gunicorn>=21,<23
